requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
DAWG-Python==0.7.2
defusedxml==0.7.1
gensim==4.3.2
idna==3.7
numpy==1.26.4
odfpy==1.4.1
pandas==2.2.2
pymorphy3==2.0.2
pymorphy3-dicts-ru==2.4.417150.4580142
pyTelegramBotAPI==4.21.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scipy==1.12.0
setuptools==70.3.0
six==1.16.0
smart-open==7.0.4
telebot==0.0.5
tzdata==2024.1
ufal.udpipe==1.3.1.1
urllib3==2.2.2
wget==3.2
wrapt==1.16.0
Flask==3.0.3